+++ src/main/java/net/MrGise/mmm/registry/compat/JEIMMMPlugin.java
@@ -6,8 +6,11 @@ import mezz.jei.api.registration.IGuiHandlerRegistration;
 import mezz.jei.api.registration.IRecipeCategoryRegistration;
 import mezz.jei.api.registration.IRecipeRegistration;
 import net.MrGise.mmm.MMM;
+import net.MrGise.mmm.compat.jei.BowyeryRecipeCategory;
 import net.MrGise.mmm.compat.jei.ThingamajigRecipeCategory;
+import net.MrGise.mmm.recipe.BowyeryRecipe;
 import net.MrGise.mmm.recipe.ThingamajigRecipe;
+import net.MrGise.mmm.screen.bowyery_table.BowyeryTableScreen;
 import net.MrGise.mmm.screen.thingamajig.ThingamajigScreen;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
@@ -25,6 +28,7 @@ public class JEIMMMPlugin implements IModPlugin {
     @Override
     public void registerCategories(IRecipeCategoryRegistration registration) {
         registration.addRecipeCategories(new ThingamajigRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
+        registration.addRecipeCategories(new BowyeryRecipeCategory(registration.getJeiHelpers().getGuiHelper()));
     }
 
     @Override
@@ -33,11 +37,17 @@ public class JEIMMMPlugin implements IModPlugin {
 
         List<ThingamajigRecipe> thingamajigRecipes = manager.getAllRecipesFor(ThingamajigRecipe.Type.INSTANCE);
         registration.addRecipes(ThingamajigRecipeCategory.THINGAMAJIG_TYPE, thingamajigRecipes);
+
+        List<BowyeryRecipe> bowyeryRecipes = manager.getAllRecipesFor(BowyeryRecipe.Type.INSTANCE);
+        registration.addRecipes(BowyeryRecipeCategory.BOWYERY_TYPE, bowyeryRecipes);
     }
 
     @Override
     public void registerGuiHandlers(IGuiHandlerRegistration registration) {
-        registration.addRecipeClickArea(ThingamajigScreen.class, 85, 32, 6, 26,
+        registration.addRecipeClickArea(ThingamajigScreen.class, 85, 36, 7, 26,
                 ThingamajigRecipeCategory.THINGAMAJIG_TYPE);
+
+        registration.addRecipeClickArea(BowyeryTableScreen.class, 84, 27, 22, 15,
+                BowyeryRecipeCategory.BOWYERY_TYPE);
     }
 }