+++ .github/logs/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.5a (skipping the -1.20.1 because it's not important)
+## [0.5a] (skipping the -1.20.1 because it's not important)
 ### Added:
 - JEI support (advancing the course)
-  - 4/4/2026, 11:15 AM: Added Thingamajig JEI support and changed some directories 
 - Changelog will be updated from now on for new versions
 - Changelog for 0.4a-passover
+
+### Commits:
+- 04/4/2026: Added Thingamajig JEI support and changed some directories
+- 04/4/2026: Added Bowyery Table JEI support and required JEI in the mods.toml file